requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
asyncio
python-binance
